drop godotenv
