--- Godeps/_workspace/src/github.com/joho/godotenv/.gitignore
@@ -1,1 +1,0 @@
-.DS_Store
