--- Godeps/_workspace/src/github.com/joho/godotenv/fixtures/equals.env
@@ -1,2 +1,0 @@
-export OPTION_A='postgres://localhost:5432/database?sslmode=disable'
-
