--- Godeps/_workspace/src/github.com/joho/godotenv/fixtures/exported.env
@@ -1,2 +1,0 @@
-export OPTION_A=2
-export OPTION_B='\n'
