--- Godeps/_workspace/src/github.com/joho/godotenv/fixtures/plain.env
@@ -1,5 +1,0 @@
-OPTION_A=1
-OPTION_B=2
-OPTION_C= 3
-OPTION_D =4
-OPTION_E = 5
