--- Godeps/_workspace/src/github.com/joho/godotenv/fixtures/quoted.env
@@ -1,8 +1,0 @@
-OPTION_A='1'
-OPTION_B='2'
-OPTION_C=''
-OPTION_D='\n'
-OPTION_E="1"
-OPTION_F="2"
-OPTION_G=""
-OPTION_H="\n"
