--- Godeps/_workspace/src/github.com/joho/godotenv/wercker.yml
@@ -1,1 +1,0 @@
-box: pjvds/golang
